% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.spFSR.R
\name{summary.spFSR}
\alias{summary.spFSR}
\title{Summarising an spFSR object}
\usage{
\method{summary}{spFSR}(object, ...)
}
\arguments{
\item{object}{A \code{spFSR} object}

\item{...}{Additional arguments}
}
\value{
Summary of an spFSR object consisting of number of features selected, wrapper type, total number of iterations, the best performing features, and the descriptive statistics of the best iteration result (the iteration where the best performing features are found).
}
\description{
Summarising an spFSR object
}
\seealso{
\link[spFSR]{getImportance}, \link[spFSR]{spFSR.default}, and \link[spFSR]{spFeatureSelection}.
}
\author{
Yong Kai Wong \email{yongkai1017@gmail.com}
}
