% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImportance.R
\name{getImportance}
\alias{getImportance}
\title{Extracting feature importance data from a spFSR object}
\usage{
getImportance(x)
}
\arguments{
\item{x}{a \code{spFSR} object}
}
\value{
A \code{data.frame} of features and feature importance
}
\description{
This returns importance ranks of best performing features. See \link[spFSR]{spFeatureSelection} for example.
}
\seealso{
\link[spFSR]{plotImportance} and \link[spFSR]{spFeatureSelection}.
}
